+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Juniper.jl Changelog
 
 ### Staged
-- nothing
+- add linear constraints using @constraint in root model
+- in BFS mode use the highest depth if several nodes have the same objective
 
 ### v0.2.4
 - Remove Julia version upper bound